fix: pass descriptor by reference in update_*_with_descriptor
This commit fixes a bug where the descriptor was being passed by value
instead of by reference in the `update_input_with_descriptor` and
`update_output_with_descriptor` methods of the `WrapPsbt` struct.

When passing by value, the `descriptor` variable was being moved into the
`match` statement. This consumed the `descriptor` variable, making it
impossible to use it again.

Issue: BTC-1743