feat: add fixture for OP_DROP (but do not enable support for it)
On the master branch, we DO want to have the OP_DROP test case, but unlike
the beta branch, we make skip the test and make sure descriptor parsing
fails with an unknown wrapper error.

This reduces the distance between `beta` and `master` branches: less
merge conflicts and easier to merge back and forth.

Issue: BTC-1829